+++ documentation/api/change_log.rst
@@ -20,6 +20,8 @@ v3.0-5 | 2022-12-30
     - ``production-price-sensor`` -> send in ``flex-context`` instead
     - ``inflexible-device-sensors`` -> send in ``flex-context`` instead
 
+- Added a subsection on deprecating and sunsetting to the Introduction section.
+- Added a subsection on describing flexibility to the Notation section.
 
 v3.0-4 | 2022-12-08
 """""""""""""""""""